CORE-17310 - Show same member in any status from member lookup API (#4788)
Adjusting the member lookup API to return the member on behalf of which the lookup is performed regardless of their status. This will allow members to always be able to check their status. The old version (v1) of the API retains the existing behaviour to ensure this change is not breaking.

Existing unit tests for v1 were copied for v5_1 and a new test case was added for the case where the member performing the lookup is suspended.