CORE-15813: Simplify VirtualNodeInfoMap
While debugging CORE-15813, I noticed this class has a few issues.

Top issue is that it holds two ConcurrentHashMap fields and attempts
to validate they are in sync but it never does any locking in its own
methods that modify these fields. Switching threads mid way through
the methods of this class could present a view of the two maps holding
different content. That would break the rules this class attempts to
enforce. To solve this I've made the methods Synchronized.

The getAllAsCordaObjects method built it's return value in two
steps. It built an in memory map and immediatly threw it away. I've
updated this method to build the final map from the start.

The put and remove methods had been broken down into methods with a
focus on which field they were handling. These methods are simple
enough to do the work in one and the resulting code makes it clear
they are updating both fields.

Extract validation logic into its own method so it is distinct from
modifications to the fields.

Remove a validation check that is impossible to hit now we have
synchronized our own methods.